+++ apps/frontend/src/modules/users/users.ts
@@ -1,21 +1,9 @@
-import { config } from "~/libs/modules/config/config.js";
-import { http } from "~/libs/modules/http/http.js";
-import { storage } from "~/libs/modules/storage/storage.js";
-
-import { UserApi } from "./users-api.js";
-
-const userApi = new UserApi({
-	baseUrl: config.ENV.API.ORIGIN_URL,
-	http,
-	storage,
-});
-
-export { userApi };
 export {
 	type UserDto,
 	type UserGetAllResponseDto,
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,
+	type UserSignUpFormDto,
 	type UserSignUpRequestDto,
 	type UserSignUpResponseDto,
 } from "./libs/types/types.js";
@@ -23,4 +11,3 @@ export {
 	userSignInValidationSchema,
 	userSignUpValidationSchema,
 } from "./libs/validation-schemas/validation-schemas.js";
-export { actions, reducer } from "./slices/users.js";