+++ apps/mobile/src/libs/enums/navigation/root-screen-name.enum.ts
@@ -1,8 +1,8 @@
 const RootScreenName = {
 	BOTTOM_TABS_NAVIGATOR: "BottomTabsNavigator",
+	QUIZ_ENTRY: "Quiz Entry",
 	SIGN_IN: "Sign In",
 	SIGN_UP: "Sign Up",
-	WELCOME: "Welcome",
 } as const;
 
 export { RootScreenName };