+++ apps/mobile/src/libs/types/navigation/root-navigation-parameter-list.type.ts
@@ -2,9 +2,9 @@ import { type RootScreenName } from "~/libs/enums/enums";
 
 type RootNavigationParameterList = {
 	[RootScreenName.BOTTOM_TABS_NAVIGATOR]: undefined;
+	[RootScreenName.QUIZ_ENTRY]: undefined;
 	[RootScreenName.SIGN_IN]: undefined;
 	[RootScreenName.SIGN_UP]: undefined;
-	[RootScreenName.WELCOME]: undefined;
 };
 
 export { type RootNavigationParameterList };