+++ packages/shared/src/modules/users/libs/types/user-sign-up-response-dto.type.ts
@@ -1,4 +1,4 @@
-import { UserDto } from "./user-dto.type.js";
+import { type UserDto } from "./user-dto.type.js";
 
 type UserSignUpResponseDto = {
 	token: string;